now loginworks for both workers and clients, ability to signup during purchase, ability to buy as a logged in custumer + fixed the way in which we handeled logins before, now details confirmation happens on the server side

+++ entities/src/main/java/il/cshaifasweng/OCSFMediatorExample/entities/CinemaMember.java
@@ -9,6 +9,7 @@ public class CinemaMember extends CasualBuyer{
 	 * 
 	 */
 	private static final long serialVersionUID = -364482658878786871L;
+	
 	private String username;
 	private String password;
 	int ticketsCredit;
@@ -16,7 +17,7 @@ public class CinemaMember extends CasualBuyer{
 	
 	public CinemaMember () {}
 	
-	public CinemaMember (String firstName, String lastName, int customerId,int creditNum, String electronicMail
+	public CinemaMember (String firstName, String lastName, int customerId, long creditNum, String electronicMail
 			, String username, String password) {
 		this.firstName = firstName;
 		this.lastName = lastName;
